Celsius only DHT library
Removed use of float types. Renamed some variables.

+++ DHT.h
@@ -30,8 +30,8 @@ class DHT {
  public:
   DHT(uint8_t pin, uint8_t type, uint8_t count=6);
   void begin(void);
-  float readTemperature(void);
-  float readHumidity(void);
+  int readTemperature(void);
+  int readHumidity(void);
   boolean read(void);
 
 };
