+++ examples/DHTtester/DHTtester.ino
@@ -41,9 +41,9 @@ void loop() {
 
   // Reading temperature or humidity takes about 250 milliseconds!
   // Sensor readings may also be up to 2 seconds 'old' (its a very slow sensor)
-  float h = dht.readHumidity();
+  int h = dht.readHumidity();
   // Read temperature as Celsius
-  float t = dht.readTemperature();
+  int t = dht.readTemperature();
   
   // Check if any reads failed and exit early (to try again).
   if (isnan(h) || isnan(t) ) {
